requirements.txt
aes-everywhere==1.2.10
alabaster==0.7.16
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
astroid==2.15.8
async-timeout==4.0.3
attrs==24.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bottle==0.12.25
bs4==0.0.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
dill==0.3.8
distro==1.8.0
Django==4.2.4
django-allauth==0.63.3
django-ckeditor==6.7.0
django-crispy-forms==1.12.0
django-js-asset==2.2.0
django-maintenance-mode==0.21.1
docutils==0.20.1
docxcompose==1.4.0
docxtpl==0.18.0
dodgy==0.2.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fastapi==0.104.1
flake8==5.0.4
flake8-polyfill==1.0.2
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
gunicorn==19.9.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
imageio==2.34.2
imagesize==1.4.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
jwt==1.3.1
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
lazy_loader==0.4
lxml==5.3.0
mando==0.7.1
MarkupSafe==2.1.5
matplotlib==3.9.1
mccabe==0.7.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
openai==1.3.3
opencv-python==4.9.0.80
openpyxl==3.1.5
orjson==3.10.1
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.3
pep8-naming==0.10.0
Pillow==10.0.0
platformdirs==4.2.2
pluggy==1.5.0
ply==3.11
prospector==1.10.3
psycopg2-binary==2.9.7
pycodestyle==2.9.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.5.1
pydantic_core==2.14.3
pydocstyle==6.3.0
pyflakes==2.5.0
pygame==2.6.0
Pygments==2.18.0
pylint==2.17.7
pylint-celery==0.3
pylint-django==2.5.3
pylint-flask==0.6
pylint-plugin-utils==0.7
pyparsing==3.1.2
PySocks==1.7.1
pytesseract==0.3.13
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-fsutil==0.14.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
radon==6.0.1
redis==5.0.1
regex==2024.5.15
requests==2.31.0
requirements-detector==1.2.2
scikit-image==0.24.0
scipy==1.14.0
selenium==4.25.0
semver==3.0.2
setoptconf-tmp==0.3.1
six==1.16.0
slimit==0.8.1
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.5
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.5.0
starlette==0.27.0
tifffile==2024.7.2
toml==0.10.2
tomlkit==0.12.5
tqdm==4.66.1
tree-sitter==0.20.4
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.2
whichcraft==0.6.1
wrapt==1.16.0
wsproto==1.2.0
XlsxWriter==3.2.0
zope.event==5.0
zope.interface==6.3
